+++ src/extension.ts
@@ -1,14 +1,14 @@
 import * as vscode from 'vscode';
 import * as _path from 'path';
 
-import { IJEEditorProvider } from './i18n-json-editor/providers/ije-editor-provider';
-import { IJEConfiguration } from './i18n-json-editor/ije-configuration';
+import { IJEEditorProvider } from './i18n-l10n-editor/providers/ije-editor-provider';
+import { IJEConfiguration } from './i18n-l10n-editor/ije-configuration';
 
 export function activate(context: vscode.ExtensionContext) {
     if (IJEConfiguration.WORKSPACE_FOLDERS) {
-        let myStatusBarItem: vscode.StatusBarItem = vscode.window.createStatusBarItem(vscode.StatusBarAlignment.Left, 100);
-        myStatusBarItem.command = 'i18n-json-editor';
-        myStatusBarItem.text = `$(symbol-string) i18n editor`;
+        let myStatusBarItem: vscode.StatusBarItem = vscode.window.createStatusBarItem(vscode.StatusBarAlignment.Left, 10);
+        myStatusBarItem.command = 'i18n-l10n-editor';
+        myStatusBarItem.text = `$(globe)$(whole-word) i18n/l10n Editor`;
         myStatusBarItem.show();
     }